+++ src/test/kotlin/mbuhot/eskotlin/query/joining/HasChildTest.kt
@@ -49,7 +49,7 @@ class HasChildTest {
                 "ignore_unmapped": false,
                 "boost": 1.0,
                 "inner_hits": {
-                    "name": "blog_tag",
+                    "ignore_unmapped": false,
                     "from": 0,
                     "size": 3,
                     "version": false,